refactor(scheduler): modularize into package with config, coordination, execution, fairness, shutdown; preserve API via shim
Break up monolithic scheduler into `src/meridian/core/scheduler/` modules and keep public imports stable. Improves readability and maintainability without behavior changes.
