+++ src/meridian/core/scheduler/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from .coordinator import Scheduler
+from .config import SchedulerConfig
+
+__all__ = ["Scheduler", "SchedulerConfig"]
